Update and rename typingGame.c to typingGame(수정).c
